+++ Example/Tests/DatabaseManipulationSpec.swift
@@ -22,19 +22,19 @@ class DatabaseManipulationSpec: SwiftyDBSpec {
             object.primaryKey = 123
             object.string = "First string"
             
-            let filters: Filter = ["primaryKey": object.primaryKey]
+            let filter: Filter = ["primaryKey": object.primaryKey]
             
             context("object is added") {
                 it("should contain the object after it is added") {
                     expect(database.addObject(object).isSuccess).to(beTrue())
-                    expect(database.dataForType(TestClass.self, matchingFilters: filters).value?.count) == 1
+                    expect(database.dataForType(TestClass.self, matchingFilter: filter).value?.count) == 1
                 }
             }
             
             context("object is deleted") {
                 it("should not contain the object after deletion") {
-                    expect(database.deleteObjectsForType(TestClass.self, matchingFilters: filters).isSuccess).to(beTrue())
-                    expect(database.dataForType(TestClass.self, matchingFilters: filters).value?.count) == 0
+                    expect(database.deleteObjectsForType(TestClass.self, matchingFilter: filter).isSuccess).to(beTrue())
+                    expect(database.dataForType(TestClass.self, matchingFilter: filter).value?.count) == 0
                 }
             }
             
@@ -45,9 +45,9 @@ class DatabaseManipulationSpec: SwiftyDBSpec {
                     object.string = "Updated string"
                     
                     expect(database.addObject(object).isSuccess).to(beTrue())
-                    expect(database.dataForType(TestClass.self, matchingFilters: filters).value?.count) == 1
+                    expect(database.dataForType(TestClass.self, matchingFilter: filter).value?.count) == 1
                     
-                    if let stringValue = database.dataForType(TestClass.self, matchingFilters: filters).value?.first?["string"] as? String {
+                    if let stringValue = database.dataForType(TestClass.self, matchingFilter: filter).value?.first?["string"] as? String {
                         expect(stringValue == "Updated string").to(beTrue())
                     }
                 }